user module updated

+++ src/app/modules/User/user.service.ts
@@ -1,5 +1,7 @@
 import { User } from '@prisma/client';
 import prisma from '../../../shared/prisma';
+import ApiError from '../../../errors/ApiError';
+import httpStatus from 'http-status';
 
 const getSingleUser = async (id: string): Promise<Partial<User> | null> => {
   const result = await prisma.user.findUnique({
@@ -17,7 +19,9 @@ const getSingleUser = async (id: string): Promise<Partial<User> | null> => {
       password: false,
     },
   });
-
+  if (!result) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'user not found!');
+  }
   return result;
 };
 
@@ -39,6 +43,14 @@ const getAllUser = async (): Promise<Partial<User>[]> => {
 };
 
 const deleteUser = async (id: string): Promise<Partial<User>> => {
+  const isUserExist = await prisma.user.findUnique({
+    where: {
+      id,
+    },
+  });
+  if (!isUserExist) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'user not found!');
+  }
   const result = await prisma.user.delete({
     where: {
       id,
@@ -62,6 +74,15 @@ const updateUser = async (
   id: string,
   payload: Partial<User>
 ): Promise<Partial<User>> => {
+  const isUserExist = await prisma.user.findUnique({
+    where: {
+      id,
+    },
+  });
+  if (!isUserExist) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'user not found!');
+  }
+
   const result = await prisma.user.update({
     where: {
       id,
